vendas/vendas-cemag/requirements.txt
Django==5.2.8
requests>=2.32.0
python-dotenv>=1.0.1
openai>=1.0.0
instructor>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
gunicorn
django-environ
psycopg2-binary
whitenoise
